requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-texttospeech==2.25.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
setuptools==78.1.0
urllib3==2.4.0
